feat(app): Back to the Future!
Revert back to Fluture package to better deal with fetch's
promise api.  In addition, handle concurrent requests to
hacker news' api and retrieve the appropriate story data.
The main request functionality is done, now I just need to
display the information in a pretty fashion.  In addition,
would be nice to refactor main request logic in a more
composable way.